+++ packages/pdfrx/lib/src/utils/native/native.dart
@@ -2,8 +2,7 @@ import 'dart:io';
 
 import 'package:flutter/services.dart';
 import 'package:path_provider/path_provider.dart';
-
-import '../../../pdfrx.dart';
+import 'package:pdfrx_engine/pdfrx_engine.dart';
 
 final isApple = Platform.isMacOS || Platform.isIOS;
 final isAndroid = Platform.isAndroid;