+++ packages/pdfrx/lib/src/widgets/pdf_widgets.dart
@@ -4,8 +4,10 @@ import 'dart:math';
 import 'dart:ui' as ui;
 
 import 'package:flutter/material.dart';
+import 'package:pdfrx_engine/pdfrx_engine.dart';
 
-import '../../pdfrx.dart';
+import '../pdf_document_ref.dart';
+import '../pdfrx_flutter.dart';
 
 /// A widget that loads PDF document.
 ///