+++ packages/pdfrx_engine/lib/src/pdf_document_event.dart
@@ -46,7 +46,7 @@ class PdfDocumentPageStatusChangedEvent implements PdfDocumentEvent {
 
   /// The pages that have changed.
   ///
-  /// The map is from page number (1-based) to it's status change.
+  /// The map is from page number (1-based) to its status change.
   ///
   /// You can assume that the keys in this map are sorted in ascending order.
   final Map<int, PdfPageStatusChange> changes;